Move `SiteEnvironmentBadge` component to components directory
