+++ client/dashboard/sites/staging-site-sync-modal/index.tsx
@@ -15,7 +15,7 @@ import { createInterpolateElement } from '@wordpress/element';
 import { __, isRTL } from '@wordpress/i18n';
 import { chevronRight, chevronLeft } from '@wordpress/icons';
 import InlineSupportLink from '../../components/inline-support-link';
-import SiteEnvironmentBadge, { EnvironmentType } from '../site-environment-badge';
+import SiteEnvironmentBadge, { EnvironmentType } from '../../components/site-environment-badge';
 
 const DirectionArrow = () => {
 	return (
